artlift_app/requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.8.1
attrs==23.2.0
cachetools==6.2.4
certifi==2024.6.2
cffi==2.0.0
charset-normalizer==3.3.2
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
deprecation==2.1.0
diagraph==0.4.7
distro==1.9.0
Django==5.1.4
django-cors-headers==4.6.0
django-filter==24.3
django-ninja==1.3.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
fire==0.7.0
frozenlist==1.4.1
fsspec==2026.1.0
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.7
inflection==0.5.1
jiter==0.10.0
Markdown==3.7
markdown-it-py==3.0.0
mdurl==0.1.2
mmh3==5.2.0
multidict==6.0.5
networkx==3.5
openai==1.101.0
packaging==24.2
postgrest==2.27.1
propcache==0.4.1
psycopg2-binary==2.9.11
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.18.0
pyiceberg==0.10.0
PyJWT==2.10.1
pyparsing==3.3.1
pyroaring==1.0.3
pyTelegramBotAPI==4.20.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
pytz==2024.2
PyYAML==6.0.2
rav==0.0.9
realtime==2.27.1
regex==2025.7.34
requests==2.32.3
rich==13.9.2
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.2
storage3==2.27.1
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.27.1
supabase-auth==2.27.1
supabase-functions==2.27.1
tenacity==9.1.2
termcolor==2.4.0
tiktoken==0.11.0
tqdm==4.67.1
twilio==9.2.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.2
websockets==15.0.1
yarl==1.22.0
